# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/export/issue/{issue_id}": {
            "get": {
                "description": "Export an issue to EPUB format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporters"
                ],
                "summary": "Export an issue",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address to send the EPUB file",
                        "name": "mail_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Issue ID",
                        "name": "issue_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Issue exported successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/export/issue/{issue_id}/article/{article_id}": {
            "get": {
                "description": "Export an issue to EPUB format",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Exporters"
                ],
                "summary": "Export an issue article",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address to send the EPUB file",
                        "name": "mail_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Issue ID",
                        "name": "issue_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Issue exported successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}